requirements.txt
langchain
langchain-openai
langchain-google-genai
fastapi
uvicorn
python-dotenv
langchain-community
pymysql
transformers
torch
unstructured
pandas
openpyxl
sentence-transformers
qdrant-client
llama-index
psycopg2-binary
streamlit
llama-index-llms-openai
hugchat
cd
faker
langchain-anthropic
langchain-qdrant
pymupdf
pydub
passlib
python-multipart
sqlalchemy
bcrypt
python-jose[cryptography]
pydantic
email-validator
pdfplumber
llama-index-llms-openai
chromadb
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index
langchain_chroma
motor
pymongo
langchain-mongodb